hw0x03/koalas-dots/app/requirements.txt
flask-wtf==0.15.1
flask-ext==0.1
flask_session==0.4.0
numpy==1.20.1